pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yamii"
version = "0.1.0"
description = "Yamii - AI Chat Service"
requires-python = ">=3.11"
dependencies = [
    "pynacl==1.5.0",
    "pydantic-settings>=2.0.0",
    "uvicorn[standard]>=0.24.0",
    "fastapi[standard]>=0.115.0",
    "aiohttp>=3.9.0",
    "websockets>=11.0.0",
    "openai>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["yamii"]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
]

dev = [
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",        # Pyflakes
    "I",        # isort
    "UP",       # pyupgrade (古い型ヒントを自動修正)
    "PIE",      # flake8-pie
    "ASYNC",    # flake8-async
]
ignore = [
    "E501",     # line too long (formatterに任せる)
    "ASYNC230", # テストでのブロッキングファイル読み込みを許容
]

[tool.ruff.format]
quote-style = "double"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.8.0",
]
